Refactor KnnVectorValues.getAcceptOrds() default to match dense assumption
